+++ CHANGELOG.md
@@ -5,13 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- [lukehsiao][lh]: add ipify.org as a fallback IP provider
 
-## [0.1.0] - 2018-08-28
+## [0.2.0] - 2019-10-02
+### Changed
+- [@lukehsiao][lh]: Switch argument parsing from clap to structopt. This
+  switches the --ipv6 from being a flag to --ip-type=<IPV4|IPV6>.
+- [@lukehsiao][lh]: Use ipify.org as a fallback IP provider for ident.me.
+
+## [0.1.0] - 2019-08-28
 Initial release on GitHub and crates.io.
 
 
 [lh]: https://github.com/lukehsiao
 
-[Unreleased]: https://github.com/lukehsiao/netlify-ddns-rs/compare/v0.1.0...master
+[Unreleased]: https://github.com/lukehsiao/netlify-ddns-rs/compare/v0.2.0...master
+[0.2.0]: https://github.com/lukehsiao/netlify-ddns-rs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/lukehsiao/netlify-ddns-rs/releases/tag/v0.1.0